EY-1404 Egen komponent for tooltip
