+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/beregne/Sammendrag.tsx
@@ -1,69 +1,19 @@
-import { BodyShort, Button, Heading, Label, Popover, Table } from '@navikt/ds-react'
+import { Heading, Table } from '@navikt/ds-react'
 import styled from 'styled-components'
-import { useRef, useState } from 'react'
-import { InformationColored } from '@navikt/ds-icons'
-import { differenceInYears, lastDayOfMonth } from 'date-fns'
+import { lastDayOfMonth } from 'date-fns'
 import { formaterDato, formaterStringDato } from '~utils/formattering'
 import { IPdlPerson } from '~shared/types/Person'
 import { Beregning } from '~shared/types/Beregning'
-
-interface ToolTipPerson {
-  fornavn: string
-  etternavn: string
-  foedselsnummer: string
-  foedselsdato: string | Date
-}
+import { GjelderTooltip } from '~components/behandling/beregne/GjelderToolTip'
 
 interface Props {
   beregning: Beregning
   soeker?: IPdlPerson
-  soesken: IPdlPerson[]
+  soesken?: IPdlPerson[]
 }
 
 export const Sammendrag = ({ beregning, soeker, soesken }: Props) => {
   const beregningsperioder = beregning.beregningsperioder
-
-  //TODO: egen komponent
-  const GjelderTooltip = ({ soesken, soeker }: { soesken: ToolTipPerson[]; soeker: ToolTipPerson }) => {
-    const [isOpen, setIsOpen] = useState(false)
-    const ref = useRef(null)
-
-    const soeskenFlokk = [...soesken, soeker]
-
-    return (
-      <>
-        {soeskenFlokk.length} barn
-        <IconButton
-          size="small"
-          ref={ref}
-          onClick={() => setIsOpen(true)}
-          icon={<InformationColored title="Få mer informasjon om beregningsgrunnlaget" />}
-        />
-        <Popover anchorEl={ref.current} open={isOpen} onClose={() => setIsOpen(false)} placement="top">
-          <PopoverContent>
-            <Heading level="1" size="small">
-              Søskenjustering
-            </Heading>
-            <BodyShort spacing>
-              <strong>§18-5</strong> En forelder død: 40% av G til første barn, 25% av G til resterende. Beløpene slås
-              sammen og fordeles likt.
-            </BodyShort>
-            <Label>Beregningen gjelder:</Label>
-            <ul>
-              {soeskenFlokk.map((soesken) => (
-                <ListWithoutBullet key={soesken.foedselsnummer}>
-                  {`${soesken.fornavn} ${soesken.etternavn} / ${soesken.foedselsnummer} / ${differenceInYears(
-                    new Date(),
-                    new Date(soesken.foedselsdato)
-                  )} år`}
-                </ListWithoutBullet>
-              ))}
-            </ul>
-          </PopoverContent>
-        </Popover>
-      </>
-    )
-  }
 
   return (
     <TableWrapper>
@@ -98,8 +48,8 @@ export const Sammendrag = ({ beregning, soeker, soesken }: Props) => {
                 {beregningsperiode.soeskenFlokk && soeker ? (
                   <GjelderTooltip
                     soesken={beregningsperiode.soeskenFlokk.map((fnr) => {
-                      const pdlPerson = soesken.find((p) => p.foedselsnummer === fnr)
-                      return pdlPerson!!
+                      const soeskenMedData = soesken?.find((p) => p.foedselsnummer === fnr)
+                      return soeskenMedData!!
                     })}
                     soeker={soeker}
                   />
@@ -127,21 +77,4 @@ export const TableWrapper = styled.div`
       max-width: 100px;
     }
   }
-`
-
-const PopoverContent = styled(Popover.Content)`
-  max-width: 500px;
-`
-
-const IconButton = styled(Button)`
-  height: 1.25rem;
-  width: 1.25rem;
-  border-radius: 50%;
-  padding: 0;
-  min-width: 0;
-  vertical-align: sub;
-  margin-left: 4px;
-`
-const ListWithoutBullet = styled.li`
-  list-style-type: none;
 `
